backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.4.0
python-dotenv==1.0.0
flask-limiter==3.5.0
